requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
celery==5.4.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
Django==4.2.13
django-celery-beat==2.6.0
django-filter==24.2
django-timezone-field==6.1.0
djangorestframework==3.15.1
flake8==7.0.0
flower==2.0.1
humanize==4.9.0
idna==3.7
install==1.3.5
kombu==5.3.7
mccabe==0.7.0
numpy==1.26.4
pandas==2.2.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
python-crontab==3.1.0
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.5
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13